# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.yourdomain.com/api/v1",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.yourdomain.com/api/v1",
      "description": "Production Environment"
    }
  ],
  "info": {
    "title": "Omni-Backend API",
    "version": "1.0.0",
    "description": "Generated by nestia - https://github.com/samchon/nestia",
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "schemas": {
...
      "object.o2": {
        "type": "object",
        "properties": {},
        "required": []
      },
...
      "PusherWebhookPayload": {
        "type": "object",
        "properties": {
          "time_ms": {
            "type": "number"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PusherWebhookEvent"
            }
          }
        },
        "required": [
          "time_ms",
          "events"
        ]
      },
      "PusherWebhookEvent": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "event": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "socket_id": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "channel",
          "event",
          "data"
        ]
      },
...
    }
  },
  "paths": {
...
    "/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "threadId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "externalMessageId": {
                        "type": "string"
                      },
                      "direction": {
                        "type": "string",
                        "enum": [
                          "inbound",
                          "outbound"
                        ]
                      },
                      "subject": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "sent",
                          "delivered",
                          "read",
                          "failed"
                        ]
                      },
                      "role": {
                        "type": "string",
                        "enum": [
                          "owner",
                          "pa"
                        ]
                      },
                      "text": {
                        "type": "string",
                        "nullable": true
                      },
                      "out": {
                        "type": "boolean",
                        "nullable": true
                      },
                      "sentAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      },
                      "threadId": {
                        "type": "string",
                        "nullable": true
                      },
                      "messageId": {
                        "type": "string",
                        "nullable": true
                      },
                      "inReplyTo": {
                        "type": "string",
                        "nullable": true
                      },
                      "references": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "nullable": true
                      },
                      "sentBy": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "role": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "enum": [
                                  "PA",
                                  "super-admin"
                                ]
                              }
                            },
                            "required": [
                              "name"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "firstName",
                          "lastName",
                          "email",
                          "role"
                        ],
                        "nullable": true
                      },
                      "provideOriginalPayload": {
                        "$ref": "#/components/schemas/object.o2"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "participants": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "id",
                      "externalMessageId",
                      "direction",
                      "status",
                      "role",
                      "sentBy",
                      "provideOriginalPayload",
                      "createdAt",
                      "updatedAt",
                      "participants"
                    ]
                  }
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "messages",
          "fetchMessages"
        ],
        "x-samchon-controller": "MessageController"
      }
    },
...
    "/webhooks/pusher/messages": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PusherWebhookPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "webhooks",
          "pusher",
          "messages",
          "handleWebhook"
        ],
        "x-samchon-controller": "PusherWebhookController"
      }
    },
...
  },
...
}